Refactor IContextualMenuClassNames to .types for export
