+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenu.base.tsx
@@ -3,7 +3,8 @@ import {
   IContextualMenuProps,
   ContextualMenuItemType,
   IContextualMenuStyleProps,
-  IContextualMenuStyles
+  IContextualMenuStyles,
+  IContextualMenuClassNames
 } from './ContextualMenu.types';
 import {
   IContextualMenuItemStyles,
@@ -39,10 +40,7 @@ import {
   ContextualMenuAnchor
 } from './ContextualMenuItemWrapper/index';
 import { IProcessedStyleSet } from '@uifabric/styling/lib/MergeStyles';
-import {
-  IMenuItemClassNames,
-  IContextualMenuClassNames
-} from 'office-ui-fabric-react/lib/components/ContextualMenu/ContextualMenu.classNames';
+import { IMenuItemClassNames } from 'office-ui-fabric-react/lib/components/ContextualMenu/ContextualMenu.classNames';
 
 const getClassNames = classNamesFunction<IContextualMenuStyleProps, IContextualMenuStyles>();
